requirements.txt
streamlit
requests
pillow
numpy
matplotlib
huggingface-hub